+++ engine/cpp/tools/runner.h
@@ -4,7 +4,7 @@
 #include <iostream>
 #include "../mc/common_mc_data.h"
 #include "savers/crystal_slice_saver.h"
-#include "savers/mol_saver.h"
+#include "savers/named_saver.h"
 #include "common.h"
 #include "error.h"
 
@@ -20,7 +20,7 @@ class Runner
     const std::string _name;
     const uint _x, _y;
     const double _totalTime, _eachTime;
-    MolSaver *_volumeSaver = nullptr;
+    NamedSaver *_volumeSaver = nullptr;
 
 public:
     static void stop();
