+++ engine/cpp/tools/savers/all_atoms_detector.cpp
@@ -1,0 +1,10 @@
+#include "all_atoms_detector.h"
+
+namespace vd {
+
+bool AllAtomsDetector::isBottom(const Atom *)
+{
+    return true;
+}
+
+}
