+++ engine/cpp/tools/savers/atom_info.h
@@ -23,8 +23,34 @@ public:
 
     const char *type() const;
     float3 coords() const;
-    std::string options() const;
+    template <class D> std::string options() const;
 };
+
+//////////////////////////////////////////////////////////////////////////////////////
+
+template <class D>
+std::string AtomInfo::options() const
+{
+    bool isBtm = D::isBottom(_atom);
+
+    int hc = _atom->hCount();
+    if (hc == 0 || isBtm)
+    {
+        hc = -1;
+    }
+
+    std::stringstream ss;
+    ss << " HCOUNT=" << hc;
+
+    ushort ac = isBtm ? _atom->valence() - _atom->bonds() : _atom->actives();
+    ac += _noBond;
+    if (ac > 0)
+    {
+        ss << " CHG=-" << ac;
+    }
+
+    return ss.str();
+}
 
 }
