+++ engine/cpp/tools/savers/mol_accumulator.h
@@ -25,6 +25,7 @@ public:
 
     void addBond(const Atom *from, const Atom *to);
 
+    template <class D>
     void writeTo(std::ostream &os, const char *prefix) const;
 
 private:
@@ -35,9 +36,46 @@ private:
     bool isNear(const Atom *first, const Atom *second) const;
 
     void writeCounts(std::ostream &os, const char *prefix) const;
+
+    template <class D>
     void writeAtoms(std::ostream &os, const char *prefix) const;
+
     void writeBonds(std::ostream &os, const char *prefix) const;
 };
+
+////////////////////////////////////////////////////////////////////////////
+
+template <class D>
+void MolAccumulator::writeTo(std::ostream &os, const char *prefix) const
+{
+    writeCounts(os, prefix);
+    writeAtoms<D>(os, prefix);
+    writeBonds(os, prefix);
+}
+
+template <class D>
+void MolAccumulator::writeAtoms(std::ostream &os, const char *prefix) const
+{
+    os << prefix << "BEGIN ATOM" << "\n";
+
+    std::vector<const AtomInfo *> orderer(_atoms.size());
+    for (auto &pr : _atoms) orderer[pr.second - 1] = &pr.first;
+
+    for (const AtomInfo *ai : orderer)
+    {
+        const float3 &coords = ai->coords();
+
+        os << prefix
+           << aiIndex(*ai) << " "
+           << ai->type() << " "
+           << coords.x << " "
+           << coords.y << " "
+           << coords.z << " "
+           << "0"
+           << ai->options<D>() << "\n";
+    }
+    os << prefix << "END ATOM" << "\n";
+}
 
 }
