+++ engine/cpp/tools/savers/named_saver.h
@@ -1,0 +1,21 @@
+#ifndef NAMED_SAVER_H
+#define NAMED_SAVER_H
+
+#include "../../atoms/atom.h"
+
+namespace vd {
+
+class NamedSaver
+{
+    std::string _name;
+
+public:
+    NamedSaver(const char *name) : _name(name) {}
+    virtual ~NamedSaver() {}
+
+    virtual void writeFrom(Atom *atom, double currentTime) = 0;
+};
+
+}
+
+#endif // NAMED_SAVER_H
