+++ engine/cpp/tools/savers/volume_saver_factory.cpp
@@ -1,13 +1,14 @@
 #include "volume_saver_factory.h"
 #include "sdf_saver.h"
+#include "surface_detector.h"
 
 namespace vd
 {
 
 VolumeSaverFactory::VolumeSaverFactory()
 {
-    registerNewType<MolSaver>("mol");
-    registerNewType<SdfSaver>("sdf");
+    registerNewType<MolSaver<SurfaceDetector>>("mol");
+    registerNewType<SdfSaver<SurfaceDetector>>("sdf");
 }
 
 }
